init_table/requirements.txt
aws-lambda-powertools==2.34.2
# temporary until new version is released:
pysqlsync @ https://github.com/hunyadi/pysqlsync/archive/refs/heads/dev.zip
instructure-dap-client[postgresql]==1.0.0
